# openapi.json
{
  "components": {
    "schemas": {
      "DetectionReason": {
        "description": "A single reason why a post might be fake.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "signal": {
            "title": "Signal",
            "type": "string"
          }
        },
        "required": [
          "category",
          "signal",
          "confidence",
          "message"
        ],
        "title": "DetectionReason",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobPost": {
        "description": "Input schema for a job posting.",
        "properties": {
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "company_website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Website"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "INR",
            "title": "Currency"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "experience_required": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Required"
          },
          "job_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Type"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "salary_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Max"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "title",
          "company_name",
          "description"
        ],
        "title": "JobPost",
        "type": "object"
      },
      "PredictionResult": {
        "description": "Output schema for fake job detection.",
        "properties": {
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/DetectionReason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "number"
          }
        },
        "required": [
          "classification",
          "risk_score",
          "confidence",
          "reasons"
        ],
        "title": "PredictionResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI-powered system to detect fraudulent job postings",
    "title": "Fake Job Posting Detector",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/batch-predict": {
      "post": {
        "description": "Detect if multiple job postings are fake.\n\nInput: List of JobPost objects\n\nOutput: List of PredictionResult objects",
        "operationId": "batch_predict_batch_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/JobPost"
                },
                "title": "Job Posts",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Predict"
      }
    },
    "/health": {
      "get": {
        "description": "Health check endpoint.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Check"
      }
    },
    "/predict": {
      "post": {
        "description": "Detect if a job posting is fake.\n\nInput:\n- title: Job title\n- company_name: Name of the company\n- description: Full job description text\n- salary_min: Minimum salary (optional)\n- salary_max: Maximum salary (optional)\n- currency: Currency code (default: INR)\n- experience_required: Experience level (e.g., \"0-2 years\")\n- job_type: Type of job (e.g., \"Full-time\")\n- platform: Source platform (e.g., \"LinkedIn\")\n- url: URL of the posting (optional)\n- company_website: Company website URL (optional)\n- location: Job location (optional)\n\nOutput:\n- classification: \"Legitimate\", \"Suspicious\", or \"Fake\"\n- risk_score: 0-100 (higher = more risk)\n- confidence: 0-100 (confidence in the classification)\n- reasons: List of specific signals/reasons for the classification",
        "operationId": "predict_job_posting_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict Job Posting"
      }
    }
  }
}